requirements.txt
openai>=1.6.1,<2.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
streamlit>=1.28.1
python-dotenv>=1.0.0
pandas>=2.1.3
numpy<2.0.0
tiktoken>=0.5.2,<0.6.0
huggingface-hub>=0.16.0
torch>=2.0.0 